Remove experimental import from testing_utils
