+++ tests/experimental/test_nash_md_trainer.py
@@ -22,7 +22,8 @@ from trl.experimental.nash_md import NashMDConfig, NashMDTrainer
 from trl.experimental.nash_md.nash_md_trainer import GeometricMixtureWrapper
 from trl.models.utils import create_reference_model
 
-from ..testing_utils import RandomPairwiseJudge, TrlTestCase, require_llm_blender, require_peft
+from ..testing_utils import TrlTestCase, require_llm_blender, require_peft
+from .testing_utils import RandomPairwiseJudge
 
 
 if is_peft_available():
