+++ tests/experimental/test_xpo_trainer.py
@@ -19,7 +19,8 @@ from transformers.utils import is_peft_available
 
 from trl.experimental.xpo import XPOConfig, XPOTrainer
 
-from ..testing_utils import RandomPairwiseJudge, TrlTestCase, require_llm_blender, require_peft
+from ..testing_utils import TrlTestCase, require_llm_blender, require_peft
+from .testing_utils import RandomPairwiseJudge
 
 
 if is_peft_available():
